Move WaitForFunding to the funding stage
