+++ src/wallet/redux/reducers/opening.ts
@@ -1,4 +1,4 @@
-import { WalletState, OpeningState, WaitForChannel, WAIT_FOR_CHANNEL, WAIT_FOR_PRE_FUND_SETUP, WAIT_FOR_FUNDING_REQUEST } from '../../states';
+import { WalletState, OpeningState, WaitForChannel, WAIT_FOR_CHANNEL, WAIT_FOR_PRE_FUND_SETUP } from '../../states';
 import { WalletAction } from '../actions';
 import { unreachable } from '../../utils';
 
@@ -7,7 +7,6 @@ export const openingReducer = (state: OpeningState, action: WalletAction): Walle
     case WAIT_FOR_CHANNEL:
       return waitForChannelReducer(state, action);
     case WAIT_FOR_PRE_FUND_SETUP:
-    case WAIT_FOR_FUNDING_REQUEST:
       return state;
     default:
       return unreachable(state);
